Added settings and systemd file to install
The settings for the ibgateway gui are now placed in the home directory
of the executing user. This is do the dev can generate a sane settings
file and copy it as the headless service won't do this by itself.
Secondly a systemd unit has been created. Lets see if that works...